+++ docs_src/pluggables/manual_self_registering.py
@@ -20,21 +20,21 @@ class MyExtension(Extension):
         # Do something here like print a log or whatever you need
         logger.success("Started the extension manually")
 
-        # Add the extension to the pluggables of Esmerald
+        # Add the extension to the extensions of Esmerald
         # And make it accessible
-        self.app.add_pluggable("my-extension", self)
+        self.app.add_extension("my-extension", self)
 
 
 @get("/home")
 async def home(request: Request) -> JSONResponse:
     """
-    Returns a list of pluggables of the system.
+    Returns a list of extensions of the system.
 
-    "pluggables": ["my-extension"]
+    "extensions": ["my-extension"]
     """
-    pluggables = list(request.app.pluggables)
+    extensions = list(request.app.extensions)
 
-    return JSONResponse({"pluggables": pluggables})
+    return JSONResponse({"extensions": extensions})
 
 
 app = Esmerald(routes=[Gateway(handler=home)])