+++ docs_src/pluggables/settings.py
@@ -15,19 +15,13 @@ my_config = PluggableConfig(name="my extension")
 
 
 class MyExtension(Extension):
-    def __init__(
-        self, app: Optional["Esmerald"] = None, config: PluggableConfig = None, **kwargs: "DictAny"
-    ):
-        super().__init__(app, **kwargs)
-        self.app = app
-
     def extend(self, config: PluggableConfig) -> None:
         logger.success(f"Successfully passed a config {config.name}")
 
 
 class AppSettings(EsmeraldAPISettings):
     @property
-    def pluggables(self) -> Dict[str, "Pluggable"]:
+    def extensions(self) -> Dict[str, Union["Extension", "Pluggable", type["Extension"]]]:
         return {"my-extension": Pluggable(MyExtension, config=my_config)}