+++ esmerald/pluggables/base.py
@@ -13,9 +13,7 @@ if TYPE_CHECKING:  # pragma: no cover
 
 class Pluggable:
     """
-    The `Pluggable` is used to create an `Esmerald` pluggable from an `Extension`.
-    When Esmerald receives pluggables, it hooks them into the system and allows
-    the access via anywhere in the application.
+    The `Pluggable` is a wrapper around an Extension to initialize it lazily.
 
     Read more about the [Pluggables](https://esmerald.dev/pluggables/) and learn how to use them.
 
@@ -43,7 +41,6 @@ class Pluggable:
             **kwargs: "DictAny",
         ):
             super().__init__(app, **kwargs)
-            self.app = app
 
         def extend(self, config: PluggableConfig) -> None:
             logger.success(f"Successfully passed a config {config.name}")
@@ -53,11 +50,11 @@ class Pluggable:
 
     pluggable = Pluggable(MyExtension, config=my_config)
 
-    app = Esmerald(routes=[], pluggables={"my-extension": pluggable})
+    app = Esmerald(routes=[], extensions={"my-extension": pluggable})
     ```
     """
 
-    def __init__(self, cls: type["Extension"], **options: Any):
+    def __init__(self, cls: type["ExtensionProtocol"], **options: Any):
         self.cls = cls
         self.options = options
 
@@ -92,8 +89,7 @@ class Extension(ABC, ExtensionProtocol):
 
     class MyExtension(Extension):
         def __init__(self, app: Optional["Esmerald"] = None, **kwargs: "DictAny"):
-            super().__init__(app, **kwargs)
-            self.app = app
+            super().__init__(app)
             self.kwargs = kwargs
 
         def extend(self, **kwargs: "DictAny") -> None:
@@ -120,12 +116,12 @@ class Extension(ABC, ExtensionProtocol):
             Doc("""Any additional kwargs needed."""),
         ],
     ):
-        super().__init__(app, **kwargs)
+        super().__init__()
         self.app = app
 
     @abstractmethod
-    def extend(self, **kwargs: "Any") -> None:
-        raise NotImplementedError("plug must be implemented by the subclasses.")
+    def extend(self, **kwargs: Any) -> None:
+        raise NotImplementedError("Extension must be implemented by the subclasses.")
 
 
 BaseExtension = Extension
@@ -135,24 +131,63 @@ class ExtensionDict(dict[str, Extension]):
     def __init__(self, data: Any = None, *, app: "Esmerald"):
         super().__init__(data)
         self.app = app
+        self.delayed_extend: Optional[dict[str, dict[str, Any]]] = {}
         for k, v in self.items():
             self[k] = v
+        while self.delayed_extend:
+            key, val = self.delayed_extend.popitem()
+            self[key].extend(**val)
+        self.delayed_extend = None
+
+    def ensure_extension(self, name: str) -> None:
+        """
+        For reordering extension initialization
+
+        class MyExtension2(Extension):
+
+            def extend(self, **kwargs: "DictAny") -> None:
+                '''
+                Function that should always be implemented when extending
+                the Extension class or a `NotImplementedError` is raised.
+                '''
+                self.app.extensions.ensure_extension("foo")
+                # Do something here
+        ```
+
+        """
+        if name not in self:
+            raise ValueError(f"Extension does not exist: {name}")
+        delayed = self.delayed_extend
+        if delayed is None:
+            return
+        val = delayed.pop(name, None)
+        if val is not None:
+            self[name].extend(**val)
 
     def __setitem__(self, name: Any, value: Any) -> None:
         if not isinstance(name, str):
-            raise ImproperlyConfigured("Pluggable names should be in string format.")
+            raise ImproperlyConfigured("Extension names should be in string format.")
         elif isinstance(value, Pluggable):
             cls, options = value
             value = cls(app=self.app, **options)
-            value.extend(**options)
+            if self.delayed_extend is None:
+                value.extend(**options)
+            else:
+                self.delayed_extend[name] = options
         elif isinstance(value, ExtensionProtocol):
-            pass
+            if self.delayed_extend is not None:
+                raise ImproperlyConfigured(
+                    "Cannot pass an initialized extension in extensions parameter."
+                )
         elif is_class_and_subclass(value, Extension):
             value = value(app=self.app)
-            value.extend()
+            if self.delayed_extend is None:
+                value.extend()
+            else:
+                self.delayed_extend[name] = {}
         else:
             raise ImproperlyConfigured(
-                "An extension must subclass from esmerald.pluggables.Extension and added to "
-                "a Pluggable object"
+                "An extension must subclass from Extension, implement the ExtensionProtocol "
+                "as instance or being wrapped in a Pluggable."
             )
         super().__setitem__(name, value)