+++ esmerald/protocols/extension.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, Optional
+from typing import Any, Optional
 
 from lilya.types import ASGIApp
 from typing_extensions import Protocol, runtime_checkable
@@ -6,6 +6,6 @@ from typing_extensions import Protocol, runtime_checkable
 
 @runtime_checkable
 class ExtensionProtocol(Protocol):  # pragma: no cover
-    def __init__(self, app: Optional["ASGIApp"] = None, **kwargs: Dict[Any, Any]): ...
+    def __init__(self, app: Optional["ASGIApp"] = None, **kwargs: Any): ...
 
     def extend(self, **kwargs: Any) -> None: ...